requirements.txt
streamlit>=1.33.0
pandas>=2.1.0
gspread>=6.0.0
google-auth>=2.27.0
bcrypt>=4.1.2
requests>=2.31.0
python-dateutil>=2.9.0